test(gui): give the remote bridge case its own file
The remote selection case was appended to service-bridge.test.ts, pushing that
file past the 700-line ceiling for test sources. Move it into its own file and
lift the fixture helpers it shares with the original file into the daemon
lifecycle helpers module, so neither file keeps a private copy.
